backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.0
anthropic==0.18.1
openai==1.12.0
newspaper3k==0.2.8
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
lxml==5.1.0
